feat: move requester to `preset-algolia`
