+++ packages/autocomplete-core/src/index.ts
@@ -1,6 +1,4 @@
 export * from './createAutocomplete';
-export * from './createRequester';
 export * from './getDefaultProps';
-export * from './requesters';
 export * from './types';
 export * from './version';
