--- packages/autocomplete-core/src/requesters/createAlgoliaRequester.ts
@@ -1,5 +1,0 @@
-import { fetchAlgoliaResults } from '@algolia/autocomplete-preset-algolia';
-
-import { createRequester } from '../createRequester';
-
-export const createAlgoliaRequester = createRequester(fetchAlgoliaResults);
