+++ packages/autocomplete-core/src/utils/__tests__/mapToAlgoliaResponse.test.ts
