+++ packages/autocomplete-core/src/utils/index.ts
@@ -4,5 +4,5 @@ export * from './getNextActiveItemId';
 export * from './getNormalizedSources';
 export * from './getActiveItem';
 export * from './isOrContainsNode';
-export * from './isRequesterDescription';
+export * from './mapToAlgoliaResponse';
 export * from './noop';
