+++ packages/autocomplete-core/src/utils/mapToAlgoliaResponse.ts
