+++ packages/autocomplete-preset-algolia/src/highlight/index.ts
@@ -1,0 +1,6 @@
+export * from './HighlightedHit';
+export * from './parseAlgoliaHitHighlight';
+export * from './parseAlgoliaHitReverseHighlight';
+export * from './parseAlgoliaHitReverseSnippet';
+export * from './parseAlgoliaHitSnippet';
+export * from './SnippetedHit';
