+++ packages/autocomplete-preset-algolia/src/index.ts
@@ -1,7 +1,3 @@
-export * from './highlight/HighlightedHit';
-export * from './highlight/parseAlgoliaHitHighlight';
-export * from './highlight/parseAlgoliaHitReverseHighlight';
-export * from './highlight/parseAlgoliaHitReverseSnippet';
-export * from './highlight/parseAlgoliaHitSnippet';
-export * from './highlight/SnippetedHit';
-export * from './search/fetchAlgoliaResults';
+export * from './highlight';
+export * from './requester';
+export * from './search';
