+++ packages/autocomplete-preset-algolia/src/requester/__tests__/getAlgoliaFacets.test.ts
