+++ packages/autocomplete-preset-algolia/src/requester/__tests__/getAlgoliaResults.test.ts
