+++ packages/autocomplete-preset-algolia/src/requester/createAlgoliaRequester.ts
@@ -1,0 +1,5 @@
+import { fetchAlgoliaResults } from '../search';
+
+import { createRequester } from './createRequester';
+
+export const createAlgoliaRequester = createRequester(fetchAlgoliaResults);
