+++ packages/autocomplete-preset-algolia/src/requester/getAlgoliaFacets.ts
