+++ packages/autocomplete-preset-algolia/src/requester/getAlgoliaResults.ts
