+++ packages/autocomplete-preset-algolia/src/requester/index.ts
@@ -1,2 +1,3 @@
+export * from './createRequester';
 export * from './getAlgoliaFacets';
 export * from './getAlgoliaResults';
