--- packages/autocomplete-preset-algolia/src/search/UserAgent.ts
@@ -1,1 +1,0 @@
-export type UserAgent = { segment: string; version?: string };
