+++ packages/autocomplete-preset-algolia/src/search/index.ts
@@ -1,0 +1,1 @@
+export * from './fetchAlgoliaResults';
